Revert "Revert "Implements MPyg123Player class""
